+++ hc/front/tests/test_details.py
@@ -56,11 +56,6 @@ class DetailsTestCase(BaseTestCase):
         r = self.client.get(self.url)
         self.assertEqual(r.status_code, 200)
 
-    def test_it_shows_new_check_notice(self):
-        self.client.login(username="alice@example.org", password="password")
-        r = self.client.get(self.url + "?new")
-        self.assertContains(r, "Your new check is ready!", status_code=200)
-
     def test_it_hides_actions_from_readonly_users(self):
         self.bobs_membership.role = "r"
         self.bobs_membership.save()