+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ### Improvements
 - Update Dockerfile to start SMTP listener (#668)
+- Implement the "Add Check" dialog
 
 ## v2.2.1 - 2022-06-13