+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Update Dockerfile to start SMTP listener (#668)
 - Implement the "Add Check" dialog
 - Include last ping type in Slack, Mattermost, Discord notifications
+- Upgrade to cron-descriptor 1.2.30
 
 ### Bug Fixes
 - Fix the display of ignored pings with non-zero exit status